requirements.txt
langchain-community
openai
aiohttp
aiofiles
beautifulsoup4
flask
python-dotenv
chromadb
